ajustes controles

+++ Artillery/Assets/_Scripts/GameManager.cs
@@ -28,13 +28,13 @@ public class GameManager : MonoBehaviour
     #endregion
 
     #region Variables
-    public int VelocidadBala { get => _VelocidadBala; set => _VelocidadBala = value; }
+    public float VelocidadBala { get => _VelocidadBala; set => _VelocidadBala = value; }
     public int DisparosPorJuego { get => _DisparosPorJuego; set => _DisparosPorJuego = value; }
     public int DisparosPorJuegoTotal { get => _DisparosPorJuegoTotal; set => _DisparosPorJuegoTotal = value; }
     public float VelocidadRotacion { get => _VelocidadRotacion; set => _VelocidadRotacion = value; }
     public float NivelSuelo { get => _NivelSuelo; set => _NivelSuelo = value; }
 
-    [SerializeField] private int _VelocidadBala = 30;
+    [SerializeField] private float _VelocidadBala = 30;
     [SerializeField] private int _DisparosPorJuego = 10;
     [SerializeField] private int _DisparosPorJuegoTotal = 10;
     [SerializeField] private float _VelocidadRotacion = 1;
